+++ crates/tabby-scheduler/src/lib.rs
@@ -1,13 +1,13 @@
 mod repository;
 
+use job_scheduler::{Job, JobScheduler};
 use std::time::Duration;
 use tracing::info;
-use job_scheduler::{JobScheduler, Job};
 
 pub fn scheduler() {
     let mut scheduler = JobScheduler::new();
 
-        // Every 5 hours.
+    // Every 5 hours.
     scheduler.add(Job::new("* * 1/5 * * *".parse().unwrap(), || {
         info!("Syncing repositories...");
         repository::sync_repositories();
