+++ crates/tabby/src/serve/mod.rs
@@ -3,7 +3,6 @@ mod completions;
 mod events;
 
 use crate::Cli;
-use tracing::info;
 use axum::{routing, Router, Server};
 use clap::{error::ErrorKind, Args, CommandFactory};
 use std::{
@@ -11,6 +10,7 @@ use std::{
     sync::Arc,
 };
 use tower_http::cors::CorsLayer;
+use tracing::info;
 use utoipa::OpenApi;
 use utoipa_swagger_ui::SwaggerUi;
